Extracted printer into menu, project, and task
