+++ lib/app.rb
@@ -1,18 +1,24 @@
+require_relative "menu_printer"
+require_relative "project_printer"
+require_relative "task_printer"
 require_relative "printer"
 require_relative "prompter"
 require_relative "input_stream"
 
 class App
-  attr_reader :printer, :input, :prompter
+  attr_reader :menu_printer, :project_printer, :task_printer, :printer, :input, :prompter
 
   def initialize(stdout, stdin)
+    @menu_printer = MenuPrinter.new(stdout)
+    @project_printer = ProjectPrinter.new(stdout)
+    @task_printer = TaskPrinter.new(stdout)
     @printer = Printer.new(stdout)
     @input = InputStream.new(stdin)
     @prompter = Prompter.new(stdout)
   end
 
   def run
-    printer.project_menu
+    menu_printer.project_menu
 
     command = input.get_project_input
 
@@ -24,16 +30,16 @@ class App
           prompter.project_name
           name = input.get_project_or_task_name
           @projects << {name => []}
-          printer.created_project(name)
+          project_printer.created_project(name)
         when "ls"
-          printer.list_of_projects
+          project_printer.list_of_projects
           if !@projects.nil? && !@projects.empty?
             @projects.each do |project|
-              printer.projects(project)
+              project_printer.projects(project)
             end
             printer.single_line
           else
-            printer.no_projects_created
+            project_printer.no_projects_created
           end
         when "d"
           if @projects && (@projects.size > 0)
@@ -41,21 +47,21 @@ class App
             project_name = input.get_project_or_task_name
             @deleted = @projects.delete_if { |project| project.keys.first == project_name.strip }.empty?
             if @deleted
-              printer.deleting_project(project_name)
+              project_printer.deleting_project(project_name)
             else
-              printer.project_does_not_exist(project_name)
+              project_printer.project_does_not_exist(project_name)
             end
           end
 
           if !@deleted && (!@projects || @projects.empty?)
-            printer.cannot_delete_project
-            printer.no_projects_created
+            project_printer.cannot_delete_project
+            project_printer.no_projects_created
           end
           @deleted = nil
         when "e"
           if !@projects || @projects.size == 0
-            printer.cannot_edit_projects
-            printer.no_projects_created
+            project_printer.cannot_edit_projects
+            project_printer.no_projects_created
             command = input.get_project_input
             next
           end
@@ -63,12 +69,12 @@ class App
           prompter.project_name
           name = input.get_project_or_task_name
           if (@current_project = @projects.detect { |p| p.keys.first == name })
-          printer.edit_project_menu(name)
+          menu_printer.edit_project_menu(name)
             command = input.get_project_input
             next
           else
-            printer.cannot_edit_project
-            printer.project_does_not_exist(name)
+            project_printer.cannot_edit_project
+            project_printer.project_does_not_exist(name)
           end
         end
       else
@@ -77,7 +83,7 @@ class App
           prompter.new_task_name
           task_name = input.get_project_or_task_name
           @current_project.values.first << task_name
-          printer.created_task(task_name)
+          task_printer.created_task(task_name)
         when "b"
           @current_project = false
           printer.double_line
@@ -87,51 +93,51 @@ class App
           new_name = input.get_project_or_task_name
           @current_project[new_name] = @current_project.values.first
           @current_project.delete(old_name)
-          printer.changed_project_name(old_name, new_name)
+          project_printer.changed_project_name(old_name, new_name)
         when "e"
           name = input.get_project_or_task_name
           if (index = @current_project.values.first.find_index(name))
-            printer.editing_task(name)
+            task_printer.editing_task(name)
             prompter.new_task_name
             new_name = input.get_project_or_task_name
             @current_project[@current_project.keys.first][index] = new_name
-            printer.changed_task_name(name, new_name)
+            task_printer.changed_task_name(name, new_name)
           else
-            printer.task_does_not_exist(name)
+            task_printer.task_does_not_exist(name)
           end
         when "d"
           project_name = @current_project.keys.first
           if @current_project.values.first.empty?
-            printer.no_tasks_created(project_name)
+            task_printer.no_tasks_created(project_name)
           else
             prompter.existing_task_name
             task_name = input.get_project_or_task_name
             if @current_project[@current_project.keys.first].delete(task_name.strip)
-              printer.deleted_task(task_name)
+              task_printer.deleted_task(task_name)
             else
-              printer.task_does_not_exist(task_name)
+              task_printer.task_does_not_exist(task_name)
             end
           end
         when "f"
           project_name = @current_project.keys.first
           if @current_project.values.first.empty?
-            printer.no_tasks_created(project_name)
+            task_printer.no_tasks_created(project_name)
           else
             prompter.existing_task_name
             task_name = input.get_project_or_task_name
             if @current_project[@current_project.keys.first].delete(task_name.strip)
-              printer.finished_task(task_name)
+              task_printer.finished_task(task_name)
             else
-              printer.task_does_not_exist(task_name)
+              task_printer.task_does_not_exist(task_name)
             end
           end
         when "ls"
           if @current_project.values.first.empty?
-            printer.no_tasks_created_in_current_project(@current_project)
+            task_printer.no_tasks_created_in_current_project(@current_project)
           else
-            printer.list_of_tasks
+            task_printer.list_of_tasks
             @current_project.values.first.each do |task|
-              printer.task(task)
+              task_printer.task(task)
             end
             printer.double_line
           end
